% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Coagulation.R
\docType{data}
\name{coagulation}
\alias{coagulation}
\title{Coagulation Data from 24 Animals}
\format{A data frame.}
\usage{
coagulation
}
\description{
Blood Coagulation of 24 animals randomly assigned to 4 different groups with different diets and sample was collected in random order.
}
\details{
\itemize{
   \item \code{coag}: Coagulation Data.
   \item \code{diet}: Diet Group.
}
}
\examples{
coagulation
}

