% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rohit_Oneway.R
\name{lsmeans.oneway}
\alias{lsmeans.oneway}
\title{Print method for the Fishers LSD.}
\usage{
lsmeans.oneway(object, ...)
}
\arguments{
\item{object}{an object of class \code{"anova"}, i.e., a fitted model.}

\item{\dots}{not used.}
}
\description{
Print method for the Fishers LSD.
}

