% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rohit_Oneway.R
\name{oneway.default}
\alias{oneway.default}
\title{Oneway Default Fucntionality}
\usage{
\method{oneway}{default}(z, ...)
}
\arguments{
\item{z}{a named list for \code{"anova"}}

\item{\dots}{not used.}
}
\description{
Oneway Default Fucntionality
}
\examples{
{
library(faraway)
data(coagulation)
RohitOnewayNewR:::oneway.default(unstack(coagulation))
}
}

