% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rohit_Oneway.R
\name{oneway.factor}
\alias{oneway.factor}
\title{Oneway Using Factors}
\usage{
\method{oneway}{factor}(z, y, ...)
}
\arguments{
\item{z}{is a Factor}

\item{y}{is values}

\item{\dots}{not used.}
}
\description{
Oneway Using Factors
}
\examples{
{
library(faraway)
data(coagulation)
z <- coagulation$diet
y <- coagulation$coag
RohitOnewayNewR:::oneway.factor(z,y)
}
}

