% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rohit_Oneway.R
\name{oneway.formula}
\alias{oneway.formula}
\title{Oneway Using Formula}
\usage{
\method{oneway}{formula}(formula, data = list(), ...)
}
\arguments{
\item{formula}{is a formula type object}

\item{data}{is a list}

\item{\dots}{not used.}
}
\description{
Oneway Using Formula
}
\examples{
{
library(faraway)
data(coagulation)
RohitOnewayNewR:::oneway.formula(as.formula(coag ~ diet),coagulation)
}
}

