% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rohit_Oneway.R
\name{print.oneway}
\alias{print.oneway}
\title{Default print for anova.}
\usage{
\method{print}{oneway}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"oneway"}, i.e., anova table.}

\item{\dots}{not used.}
}
\description{
Default print for anova.
}
\examples{
{
library(faraway)
data(coagulation)
DefaultData <- RohitOnewayNewR:::oneway.default(unstack(coagulation))
RohitOnewayNewR:::print.oneway(DefaultData)
}
}

