% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rohit_Oneway.R
\name{print.summary.oneway}
\alias{print.summary.oneway}
\title{Print method for the summary method.}
\usage{
\method{print}{summary.oneway}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"summary.anova"}, i.e., a fitted model.}

\item{\dots}{not used.}
}
\description{
Print method for the summary method.
}
\examples{
{
library(faraway)
data(coagulation)
DefaultData <- RohitOnewayNewR:::oneway.default(unstack(coagulation))
SummaryObj<- RohitOnewayNewR:::summary.oneway(DefaultData)
RohitOnewayNewR:::print.summary.oneway(SummaryObj)
}
}

