% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rohit_Oneway.R
\name{summary.oneway}
\alias{summary.oneway}
\title{Oneway Summary}
\usage{
\method{summary}{oneway}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"oneway"}, i.e., anova table.}

\item{\dots}{not used.}
}
\description{
Oneway Summary
}
\examples{
{
library(faraway)
data(coagulation)
DefaultData <- RohitOnewayNewR:::oneway.default(unstack(coagulation))
RohitOnewayNewR:::summary.oneway(DefaultData)
}
}

