% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rohit_Oneway.R
\name{oneway}
\alias{oneway}
\title{Oneway: An Alternative for Anova.}
\usage{
oneway(z, ...)
}
\arguments{
\item{z}{A list of factors}

\item{y}{A list of values}

\item{z}{a named list for \code{"anova"}}

\item{\dots}{not used.}
}
\value{
object an Object of class \code{oneway}, basically a list of all elements:
 \item{x}{ A list of values }

 \item{y}{ A list of Factors }

 \item{SSB}{ Sum of Squares Between }

 \item{SSW}{ Sum of Squares Within }

 \item{Degree_of_freedomB}{ Degrees of Freedom Between }

 \item{Degree_of_freedomW}{ Degrees of Freedom Within }

 \item{Total_mean}{ Total Mean of all elements }

 \item{Group_mean_vector}{ A list of mean of each group }

 \item{Group_length_vector}{ A list of length of each group }

 \item{Group_var_vector}{ A list of Variance of each Group }

 \item{names_groups}{ Names of all groups }

 \item{F_value}{ F value }
 Oneway Object
}
\description{
Oneway: An Alternative for Anova.
}

